Revert "replace double with Float to fix build errors"
This reverts commit 145d2662a17c1e2ba66012b3e91a5d075382796e.

+++ libhussar/include/hussar/integrators/path.h
@@ -459,9 +459,9 @@ protected:
     long sampleIndexOffset;
 
 #ifdef __CUDACC__
-    Float totalWeight;
+    double totalWeight;
 #else
-    std::atomic<Float> totalWeight;
+    std::atomic<double> totalWeight;
 #endif
 
     HUSSAR_CPU_GPU void incrementTotalWeight(Float sampleWeight) {
@@ -470,7 +470,7 @@ protected:
 #else
         /// @todo this is not elegant
         // (and could be more efficient under GCC)
-        Float oldV = totalWeight;
+        double oldV = totalWeight;
         while (!totalWeight.compare_exchange_weak(oldV, oldV + sampleWeight));
 #endif
     }
